Refactored code

+++ audio_diarization.py
@@ -5,7 +5,6 @@ import config
 import ffmpy 
 import glob
 import shutil
-import re
 
 
 def read_file(filename:str) -> pd.DataFrame:
@@ -62,7 +61,7 @@ def file_conversion(file_list: list, input_d:str, wav_dir:str)-> None:
     if (os.path.exists(wav_dir)) != True:
         for file in file_list:
             ff = ffmpy.FFmpeg(inputs={input_d + "/" +file: None}, 
-                                outputs={re.sub(r"\s+", "", file[:-4] + ".wav"): None})
+                                outputs={file[:-4] + ".wav": None})
             ff.run()
 
         files_to_move = glob.glob("*.wav")
@@ -72,7 +71,7 @@ def file_conversion(file_list: list, input_d:str, wav_dir:str)-> None:
             shutil.move(file, new_path)
 
 
-def diratzation(file_list: list, input_d: str) -> None:
+def diratzation(file_list: list, input_d: str, rttm_dir: str) -> None:
     """
     extracts subtitles and tagged speakers from
     provided audio files
@@ -85,36 +84,20 @@ def diratzation(file_list: list, input_d: str) -> None:
     None
 
     """
+    if (os.path.exists(rttm_dir)) != True:
+        from pyannote.audio import Pipeline
 
-    from pyannote.audio import Pipeline
+        pipeline = Pipeline.from_pretrained("pyannote/speaker-diarization@2.1",
+                                            use_auth_token=config.auth_token)
 
-    pipeline = Pipeline.from_pretrained("pyannote/speaker-diarization@2.1",
-                                        use_auth_token=config.auth_token)
-
-    
-    
-    full_files = [input_d + "/" + filename for filename in file_list]
-    diar_func = lambda x: pipeline(x)
-    diarizations = [diar_func(file) for file in full_files]
-    rttm_dir = "audiorttm"
-    os.mkdir(rttm_dir)
-    rttm_files = [rttm_dir + "/" + filename[:-4] + ".rttm" for filename in file_list]
-    
-    for file, diar in zip(rttm_files, diarizations):
-        with open(file, "w") as rttm:
-            diar.write_rttm(rttm)
-
-
-links = read_file("links.csv")
-audio_file_dir = "audio"
-scrape_audio(scrape_links=links, 
-            link_col="Link", 
-            output_d=audio_file_dir)
-
-audio_files = os.listdir(audio_file_dir)
-
-audio_wav_dir = "audiowav"
-file_conversion(file_list = audio_files, 
-                input_d="audio", wav_dir=audio_wav_dir)
-wav_files = os.listdir(audio_wav_dir)
-diratzation(file_list=wav_files, input_d=audio_wav_dir)
+        
+        
+        full_files = [input_d + "/" + filename for filename in file_list]
+        diar_func = lambda x: pipeline(x)
+        diarizations = [diar_func(file) for file in full_files]
+        os.mkdir(rttm_dir)
+        rttm_files = [rttm_dir + "/" + filename[:-4] + ".rttm" for filename in file_list]
+        
+        for file, diar in zip(rttm_files, diarizations):
+            with open(file, "w") as rttm:
+                diar.write_rttm(rttm)
